Refactoring for readability.
Moved EndpointDiscoveryTests to separate file.
Moved MockResponse to tests.__init__
