+++ mentions/tests/__init__.py
@@ -18,3 +18,19 @@ class WebmentionTestCase(TestCase):
             SimpleMention,
         ]:
             Model.objects.all().delete()
+
+
+class MockResponse:
+    """Mock of requests.Response."""
+
+    def __init__(
+        self,
+        url: str,
+        headers: dict = None,
+        text: str = None,
+        status_code: int = None,
+    ):
+        self.url = url
+        self.text = text
+        self.status_code = status_code
+        self.headers = headers
